backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
python-multipart
firebase-admin
